requirements.txt
annotated-types==0.6.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.3
psycopg==3.1.16
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dotenv==1.0.0
smmap==5.0.1
SQLAlchemy==2.0.23
typing_extensions==4.9.0
